Upgrade to SpringBoot 3.1.3 - Clean up dependencies - Create module for error-handling - Clean up gradle builds
