+++ app/build.gradle
@@ -13,8 +13,7 @@ dependencies {
     implementation platform("io.opentelemetry:opentelemetry-bom:${opentelemetry_version}")
     implementation "io.opentelemetry:opentelemetry-api:${opentelemetry_version}"
     runtimeOnly "io.opentelemetry.javaagent:opentelemetry-javaagent:${opentelemetry_version}"
-    runtimeOnly "org.springframework.boot:spring-boot-devtools"
-    testImplementation "io.projectreactor:reactor-test"
+
 }
 
 bootJar {
