+++ app/cache/build.gradle
@@ -4,5 +4,5 @@ dependencies {
 
     implementation "org.springframework.boot:spring-boot"
     implementation "org.springframework.boot:spring-boot-autoconfigure"
-    implementation("org.springframework.boot:spring-boot-starter-data-redis")
+    implementation "org.springframework.boot:spring-boot-starter-data-redis"
 }
