+++ app/models/build.gradle
@@ -1,3 +1,3 @@
 dependencies {
-    implementation "jakarta.validation:jakarta.validation-api:${validation_jakarta_version}"
+    implementation "jakarta.validation:jakarta.validation-api"
 }
