+++ app/persistence/build.gradle
@@ -3,7 +3,6 @@ dependencies {
 
     implementation "org.springframework.boot:spring-boot-starter-data-mongodb"
 
-    testImplementation "org.testcontainers:mysql:${test_containers_version}"
-    testImplementation("org.testcontainers:testcontainers")
-    testImplementation("org.springframework.boot:spring-boot-starter-test")
+    testImplementation "org.testcontainers:testcontainers"
+    testImplementation "org.springframework.boot:spring-boot-starter-test"
 }
