+++ app/rest/build.gradle
@@ -3,8 +3,6 @@ dependencies {
     implementation project(":app:models")
 
     implementation "org.springdoc:springdoc-openapi-starter-webmvc-ui:${springdoc_openapi_version}"
-    implementation "org.springframework.boot:spring-boot-starter-web"
-    implementation "jakarta.validation:jakarta.validation-api:${validation_jakarta_version}"
 
-    implementation "io.micrometer:micrometer-registry-prometheus:${metrics_prometheus_version}"
+    implementation "io.micrometer:micrometer-registry-prometheus"
 }
