+++ app/service/build.gradle
@@ -6,6 +6,6 @@ dependencies {
     implementation project(":app:persistence")
 
     implementation "org.springframework.boot:spring-boot-starter-data-mongodb"
-    implementation "org.springframework.boot:spring-boot-starter-web"
-    implementation "jakarta.validation:jakarta.validation-api:${validation_jakarta_version}"
+    implementation "org.springframework.boot:spring-context"
+    implementation "com.fasterxml.jackson.datatype:jackson-datatype-jsr310"
 }
