+++ app/service/src/main/java/com/simple/service/converter/SimpleKafkaConverterImpl.java
@@ -1,7 +1,6 @@
 package com.simple.service.converter;
 
 import java.io.IOException;
-import java.util.Objects;
 
 import org.springframework.stereotype.Component;
